We check whether the maintainance file exists before deleting it.

+++ app/helpers/UpdateHelper.php
@@ -262,6 +262,8 @@ class UpdateHelper
             UpdateHelper::updateAppPost($updateVersion);
         }
 
-        unlink($maintainanceFile);
+        if (file_exists($maintainanceFile)) {
+            unlink($maintainanceFile);
+        }
     }
 }
